+++ src/MyWebLog.Data/PostgreSql/PostgreSqlUploadData.fs
@@ -9,12 +9,12 @@ open Npgsql.FSharp
 type PostgreSqlUploadData (conn : NpgsqlConnection) =
 
     /// The INSERT statement for an uploaded file
-    let upInsert = """
-        INSERT INTO upload (
+    let upInsert =
+        "INSERT INTO upload (
             id, web_log_id, path, updated_on, data
         ) VALUES (
             @id, @webLogId, @path, @updatedOn, @data
-        )"""
+        )"
     
     /// Parameters for adding an uploaded file
     let upParams (upload : Upload) = [
@@ -38,31 +38,29 @@ type PostgreSqlUploadData (conn : NpgsqlConnection) =
     /// Delete an uploaded file by its ID
     let delete uploadId webLogId = backgroundTask {
         let theParams = [ "@id", Sql.string (UploadId.toString uploadId); webLogIdParam webLogId ]
-        let! tryPath =
+        let! path =
             Sql.existingConnection conn
             |> Sql.query "SELECT path FROM upload WHERE id = @id AND web_log_id = @webLogId"
             |> Sql.parameters theParams
             |> Sql.executeAsync (fun row -> row.string "path")
-        match List.tryHead tryPath with
-        | Some path ->
+            |> tryHead
+        if Option.isSome path then
             let! _ =
                 Sql.existingConnection conn
                 |> Sql.query "DELETE FROM upload WHERE id = @id AND web_log_id = @webLogId"
                 |> Sql.parameters theParams
                 |> Sql.executeNonQueryAsync
-            return Ok path
-        | None -> return Error $"""Upload ID {UploadId.toString uploadId} not found"""
+            return Ok path.Value
+        else return Error $"""Upload ID {UploadId.toString uploadId} not found"""
     }
     
     /// Find an uploaded file by its path for the given web log
-    let findByPath (path : string) webLogId = backgroundTask {
-        let! upload =
-            Sql.existingConnection conn
-            |> Sql.query "SELECT * FROM upload WHERE web_log_id = @webLogId AND path = @path"
-            |> Sql.parameters [ webLogIdParam webLogId; "@path", Sql.string path ]
-            |> Sql.executeAsync (Map.toUpload true)
-        return List.tryHead upload
-    }
+    let findByPath path webLogId =
+        Sql.existingConnection conn
+        |> Sql.query "SELECT * FROM upload WHERE web_log_id = @webLogId AND path = @path"
+        |> Sql.parameters [ webLogIdParam webLogId; "@path", Sql.string path ]
+        |> Sql.executeAsync (Map.toUpload true)
+        |> tryHead
     
     /// Find all uploaded files for the given web log (excludes data)
     let findByWebLog webLogId =
